requirements.txt
lpips==0.1.4
matplotlib==3.7.2
numpy==1.24.3
opencv_python==4.5.3.56
packaging==23.1
Pillow==10.0.0
Pillow==9.4.0
Pillow==10.0.1
PyYAML==5.4.1
PyYAML==6.0.1
scikit_image==0.19.3
scipy==1.9.3
torch==1.8.1+cu111
torchvision==0.9.1+cu111
tqdm==4.62.0
mamba_ssm==1.0.1
causal_conv1d==1.0.0
nibabel>=5.2.0
